fps fix

+++ simpleclient-1.8.9/src/main/java/simpleclient/mixin/feature/zoom/GameRendererMixin.java
@@ -11,6 +11,6 @@ import simpleclient.feature.Zoom;
 public abstract class GameRendererMixin {
     @Inject(at = @At(value = "RETURN"), method = "getFov", cancellable = true)
     public void getFov(float partialTicks, boolean changingFov, CallbackInfoReturnable<Double> cir) {
-        if (Zoom.ACTIVE) cir.setReturnValue(cir.getReturnValue() * (1 - Zoom.ZOOM_FACTOR));
+        if (Zoom.ACTIVE) cir.setReturnValue((double) (cir.getReturnValue() * (1 - Zoom.ZOOM_FACTOR)));
     }
 }
